Major clean up across several files
Mainly added new features and scenarios to support the approval and ownership types of Requests in the Curators Worklist.
File 357-curator_opens_my_worklist is deleted as all scenarios are replaced (by my_worklist.feature) or moved to my_worklist_serach.
Have moved several DOI-related scenarios to doi.feature, a start to review them as part of the worklist.

A collateral happening; I moved several log-in related scenarios from mvp.feature to user-logs-in.feature - mvp.feature needs to be removed soon.

Jira is not updated yet, tags like "needJiraTag", "updated" and comments are used to remember this - too after vacation.