Np 48502 channel ownership (#291)
Initial set of Gherkins for channel ownership.

Institutions should be able to claim channels from the Channel Registry
as their own. With this claim follows a set of constraints, within a
scope, dictating who is allowed to do what with metadata when publishing
on the channel.
In the first delivery, the focus is on claiming the channels, and thus
the constraints should be open (which means they are the same as they
are today), and the scope should be set to instance types DegreePhd,
DegreeMaster, and DegreeBachelor.

In addition, the ticket system has to be tweaked so that when publishing
on a claimed channel, it is curators at the channel owner that receives
the ticket, not curators at the Registrators institution.

---------

Co-authored-by: Joachim Jørgensen <joachim.jorgensen@Joachims-MacBook-Pro.local>